Remove stopgap interpreter.
We are now using stack reconstruction to recover from a guard failure, so
the stopgap interpeter is no longer needed.
