+++ ykfr/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "yksgi"
+name = "ykfr"
 version = "0.1.0"
 authors = ["The Yk Developers"]
 edition = "2018"
